Fixed a number of bugs associated with region mode. The middle and right mouse now ignore the region setting. The sliding of a face never creates a slab of 0 thickness. The color of the selected face is more obvious. The face gets unselected as soon as the mouse is released. The rendering immediately reverts to correct numTransforms as soon as the mouse is released.

+++ apps/vaporgui/session.cpp
@@ -243,7 +243,7 @@ void Session::resetCommandQueue(){
 // This data is inserted into currentDatastatus;
 DataStatus* Session::
 setupDataStatus(){
-	int numTimeSteps = currentMetadata->GetNumTimeSteps()[0];
+	size_t numTimeSteps = currentMetadata->GetNumTimeSteps()[0];
 	int numVariables = currentMetadata->GetVariableNames().size();
 	DataStatus* ds = new DataStatus(numVariables, numTimeSteps);
 	int numXForms = currentMetadata->GetNumTransforms();
@@ -254,7 +254,7 @@ setupDataStatus(){
 	//As we go through the variables and timesteps, keepTrack of min and max times
 	minTimeStep = 1000000000;
 	maxTimeStep = 0;
-	for (size_t ts = 0; ts<(size_t)numTimeSteps; ts++){
+	for (size_t ts = 0; ts< numTimeSteps; ts++){
 		for (int var = 0; var< numVariables; var++){
 			//Find the minimum number of transforms available on disk
 			//Start at the max (lowest res) and move down to min
